Pages are responsible for adding to the Route object thier respective routes.
Rather than having a large dictionary called beforehand, this allows us to call
an adding function inside the file that the page is being programmed in. It
means that there isn't a global object that is making routes. Instead each
route is responsible for adding itself to the global route.