pathing for static files
info:
  url_for is a file that is surved by flask backend.
  a standard path 'static/images/foo.png' is served by apache

We want our js to request files like our landing page image, so we won't be
using flask to get those types of files. It is also faster to use apache.

url_for is more for development, when your folders are changing, but since we
need to go the other direction for js, we can just have it reference the
`static` folder.

mithril is now finding static items, and apache will still be able to import
static items.