Removed __author__ and related tags from python files.
They were scattered in some, but not all python files in the project. So this
takes them out, and allows us to put them in, in a unified format later.
We can add in some of these tags in a standardized format before publishing.

If you want to know who first created a file you can use the git repo.