Fixed up genome_example.py and added that functionality to a Flask endpoint.
This isn't a permanent spot, but a stepping stone to getting genome uploading
to work for future commits.

Make sure you aren't running this with a database that matters in the
background, as this will add bad data to it. I don't think there are
protections atm for running 'test' to set a different BG server.