Endpoint downloads tab separated txt file for a sample sparql query
Moved makecsv response to the Response class
Changed it to take in a sparql response object instead of a json'ed sparql response object
Added another method that does the above, and sends the tsv back as a file.
Changed makecsv to not put tabs at the start of a line
Changed makecsv to be more pythonic (See below)
   instead of appending to a string in every operation, we make a list that is joined together.
   the function is no longer responsible for setting the accession key at the start of the tsv
      Sparql queries state the order of variables to be listed, so head.vars is listed in the order
          that the query is asked.
Changed the sample query to ask for accession first.
Changed the sample query to not ask for the internal genome uri. (Not usefull to users at this time.)